requirements.txt
fastapi
uvicorn
openai
azure-core
azure-ai-formrecognizer
azure-storage-blob
PyPDF2
pdfplumber
langchain
tabulate
docx
nltk
scikit-learn
langchain-community
langchain_openai
langchain_experimental
msal
pandas
langchain_core
numpy
mysql-connector-python
requests
pdfkit
python-multipart
pydantic
gunicorn
python-docx
pypdf2
reportlab
pycryptodome
openpyxl  
langchain[all]  